Add comparison of d_ino and st_ino for debugging
Explanation -- when I enter relative paths outside of the current
working directory as arguments to my ls program, I notice that
everything returned by 'stat()' seems wrong. It looks like the inodes
don't even match.

I suspect this is because the first argument to 'stat()' is meant to be
a full path, and I am giving it just the filename.

It seems like there are maybe a couple of ways to fix this -- I could
somehow get the whole path, or I could somehow get a file descriptor for
each file and use fstat.

A file descriptor might be best. I suspect using the path could prove
brittle.